requirements.txt
# Core dependencies
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
PyYAML>=6.0,<7.0
python-dotenv>=1.0.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
joblib>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0

# Data fetching
requests>=2.28.0,<3.0.0
yfinance>=0.2.30,<1.0.0
lxml>=4.9.0,<6.0.0
feedparser>=6.0.0,<7.0.0

# NLP / sentiment
vaderSentiment>=3.3.2,<4.0.0

# Optional: regime detection
hmmlearn>=0.3.0,<1.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0
